split into different components
